Conflicts:
	app/static/js-templates/discovery.html
	app/static/js-templates/index.html

+++ app/views.py
@@ -36,7 +36,7 @@ def profile_id(user_id):
 
 
 def unix_time(dt):
-    epoch = datetime.datetime.utcfromtimestamp(0)
+    epoch = datetime.datetime.fromtimestamp(0)
     delta = dt - epoch
     return delta.total_seconds()
 
@@ -96,7 +96,6 @@ def dummy():
     db.session.commit()
 
     return "Dummy data successfully created/verified"
-
 
 def check_valid_new_event(form):
     str = ""
@@ -179,7 +178,16 @@ def event(event_id):
     elif request.method == "PUT":
         for key in request.form:
             if key in e.__dict__:
-                setattr(e, key, request.form[key])
+                if key == "start_time":
+                    start_time = parser.parse(request.form['start_time'])
+                    setattr(e, "start_time", start_time)
+                elif key == "end_time":
+                    end_time = parser.parse(request.form['end_time'])
+                    setattr(e, "end_time", end_time)
+                elif key =="is_private":
+                    setattr(e, "is_private", True)
+                elif key == "name" or key == "description":
+                    setattr(e, key, request.form[key])
             elif key == "user_list":
                 try:
                     L = json.loads(request.form[key])
@@ -197,6 +205,8 @@ def event(event_id):
         return json_out({"status_code": 0})
 
     elif request.method == "DELETE":
+        if e.manager_id != current_user.id:
+            return json_out({"status_code": 1})
         db.session.delete(e)
         db.session.commit()
         return json_out({"status_code": 0})
@@ -296,7 +306,14 @@ def task(task_id):
     elif request.method == "PUT":
         for key in request.form:
             if key in t.__dict__:
-                setattr(t, key, request.form[key])
+                if key == "start_time":
+                    start_time = parser.parse(request.form['start_time'])
+                    setattr(t, "start_time", start_time)
+                elif key == "end_time":
+                    end_time = parser.parse(request.form['end_time'])
+                    setattr(t, "end_time", end_time)
+                elif key == "name" or key == "description" or key == "location":
+                    setattr(t, key, request.form[key])
             elif key == "user_list":
                 try:
                     L = json.loads(request.form[key])
@@ -314,6 +331,8 @@ def task(task_id):
         return json_out({"status_code": 0})
 
     elif request.method == "DELETE":
+        if t.event and t.event.manager_id != current_user.id:
+            return json_out({"status_code": 1})
         db.session.delete(t)
         db.session.commit()
         return json_out({"status_code": 0})
